apitft/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
black==23.12.0
click==8.1.7
Django==5.0
django-rest-registration==0.8.3
djangorestframework==3.14.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
pytz==2023.3.post1
sqlparse==0.4.4
yarl==1.9.4
